Feature ETP-4031: Extract shared PDF helpers to pdfUtils.js
Extract fetchJson, fetchAll, fetchOptionalJson, fetchLocationAddress,
blobToDataUrl, fetchImageDataUrl, renderPdf, and COMMON_HANDLEBARS_HELPERS
into tools/app-shell/src/windows/custom/shared/pdfUtils.js.
Both useShipmentPdf.js and useInvoicePdf.js now import from pdfUtils,
eliminating the Sonar duplication between the two files.
